successful use of API / other bugs
the API runs successfully however I wanted to add a little extra touch to practice CSS a bit. working on creating an eventListener for 'click' on the image and then having the image expand to a larger size. on any other 'click' the  enlarged img will go away and the screen will return to prev